_app.tsxとgameの切り分け
